requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Bcrypt==0.7.1
gunicorn==20.0.4
PyJWT==2.0.1
psycopg2-binary==2.8.6
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-mock==3.5.1
pytest-recording==0.11.0
SQLAlchemy==1.3.19
uuid==1.30
Werkzeug==1.0.1
boto3==1.15
uuid==1.30
